workers/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
google-cloud-firestore>=2.16.0
google-cloud-tasks>=2.16.0
google-genai>=1.0.0
pydantic>=2.7.0
pandas>=2.2.0
matplotlib>=3.9.0
requests>=2.32.0
